api/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
databases[aiomysql]==0.7.0
mysql-connector-python==8.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0